Refactor headers and adds LFO engine with waveform generation and user-defined shapes
Implements a high-performance wavetable-based LFO engine supporting sine, triangle, sawtooth, square, and user-defined waveforms. Introduces user-editable modulation shapes with curvature handling and real-time processing. Refactors filter utilities into a dedicated file and updates parameter management for improved modularity and clarity.
